Created seperate seeding module
