+++ dijkstra/run_blastseeds.py
@@ -1,6 +1,7 @@
 import argparse
 #from research_for_graph import *
 from graph_research import *
+import seeding
 import lzma
 import time
 import os.path
@@ -21,45 +22,6 @@ def initParser():
     return parser
 
 
-def adj_mat(nodes, graph):
-    print('nodes:',nodes)
-    mat = [[0]*len(nodes) for _ in range(len(nodes))]
-    for i in range(len(mat)):
-        for j in range(i, len(mat)):
-            if graph.has_edge(graph.indexes[nodes[i]],graph.indexes[nodes[j]]):
-                mat[i][j] = 1
-                mat[j][i] = 1
-    return mat
-
-
-
-def get_g1_seed(g1_seed_file) -> dict:
-    g1_seed = defaultdict(list)
-    for line in open(g1_seed_file):
-        line = line.split()
-        g1_seed[line[0]].append(line[1:])
-    return g1_seed
-
-
-
-
-def generate_seed(g1_seed_file, g2_seed_file):
-    g1_seed = get_g1_seed(g1_seed_file)
-    #print (g1_seed)
-    for line in open(g2_seed_file):
-        line = line.split()
-        if line[0] in g1_seed:
-            for nodes in g1_seed[line[0]]:
-                yield nodes,line[1:]
-
-        
-
-
-def get_seed_length(network) -> int:
-    #RNorvegicus_5_30_300000_MAX.txt 
-    info = network.split("/")
-    data = info[-1].split("_")
-    return data[1]
 
 def write_log(fname, runtime, seed, delta):
     logname = fname.replace("dijkstra", "log")
@@ -86,41 +48,26 @@ if __name__ == '__main__':
     g1_seed_file = args.g1seed
     g2_seed_file = args.g2seed
     
-    #print (len(graph1), len(graph2))
 
-    seed_length = get_seed_length(g1_seed_file)
-    #print("seed length")
-    #print(seed_length)
+    seed_length = seeding.get_seed_length(g1_seed_file)
 
-    #sims = get_degree_diff_matrix(graph1, graph2)
     sims = get_sim(args.sim, graph1, graph2, args.pickle)
 
-    for seed in generate_seed(g1_seed_file,g2_seed_file):
+    for seed in seeding.generate_seed(g1_seed_file,g2_seed_file):
         
         n1, n2 = seed
-        mat1 = adj_mat(n1,graph1)
-        mat2 = adj_mat(n2,graph2)
+        mat1 = seeding.adj_mat(n1,graph1)
+        mat2 = seeding.adj_mat(n2,graph2)
         if mat1 != mat2:
             print(mat1)
             print(mat2)
 
         start = time.time()
-        #a, b, pairs = dijkstra(graph1, graph2, get_aligned_seed(zip(*seed), graph1, graph2), sims, delta)
         a, b, pairs = stop_align2(graph1, graph2, get_aligned_seed(zip(*seed),graph1, graph2), sims, delta)
         subgraph = induced_subgraph(graph1, graph2, list(pairs))
         cov = coverage(graph1, graph2, subgraph)[0]
         cov2 = coverage(graph1, graph2, subgraph)[1]
-#        if len(pairs) == 4:
-#            print(subgraph)
-#            print(pairs)
-#            pairs = list(pairs)
-#            for i in range(len(pairs)):
-#                for j in range(i+1,len(pairs)):
-#                    print(pairs[i],pairs[j])
-#                    print(graph1.has_edge(pairs[i][0], pairs[j][0]),graph2.has_edge(pairs[i][1], pairs[j][1]) )
-#            
 
-        #print(cov)
         print("Edges aligned in subgraph:",len(subgraph))
         print("Pairs aligned in subgraph:",len(pairs))
         newcov = round(cov, 2)
